requirements.txt
Flask
Pygments
Pillow
Gunicorn